refactor before:browser:launch handling into own file
